add minimalistic mock for yarn node
