--- granulate_utils/config_feeder/client/__init__.py
@@ -1,13 +1,0 @@
-from granulate_utils.config_feeder.client.bigdata import get_node_info
-from granulate_utils.config_feeder.client.client import ConfigFeederClient
-from granulate_utils.config_feeder.client.exceptions import APIError, ClientError, MaximumRetriesExceeded
-from granulate_utils.config_feeder.client.http_client import AuthCredentials
-
-__all__ = [
-    "AuthCredentials",
-    "ConfigFeederClient",
-    "ClientError",
-    "APIError",
-    "MaximumRetriesExceeded",
-    "get_node_info",
-]
