--- granulate_utils/config_feeder/client/spark/utils.py
@@ -1,1 +1,0 @@
-SPARK_HISTORY_SERVER_DEFAULT_ADDRESS = "http://localhost:18080"
