--- granulate_utils/config_feeder/client/yarn/models.py
@@ -1,7 +1,0 @@
-from typing import Any, Dict
-
-from pydantic import BaseModel
-
-
-class YarnConfig(BaseModel):
-    config: Dict[str, Any]
