--- granulate_utils/config_feeder/core/models/aggregation.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-from granulate_utils.config_feeder.core.models.collection import CollectorType
-
-
-class NodeResourceConfigCreate(BaseModel):
-    collector_type: CollectorType
-    config_json: str
