--- tests/granulate_utils/config_feeder/fixtures/base.py
@@ -1,137 +1,0 @@
-from __future__ import annotations
-
-from abc import ABC
-from typing import Any, Callable, ContextManager, Dict, List, Tuple
-from unittest.mock import Mock, mock_open, patch
-
-from requests_mock.mocker import Mocker
-
-from granulate_utils.config_feeder.client.bigdata import get_node_info
-from granulate_utils.config_feeder.core.models.node import NodeInfo
-
-
-class NodeMockBase(ABC):
-    def __init__(self) -> None:
-        self._files: Dict[str, str] = {}
-        self._stdout: Dict[str, bytes | str] = {}
-        self._requests: List[Tuple[str, str, Dict[str, Any]]] = []
-        self._contexts: List[Tuple[ContextManager[Any], Callable[[Any], None] | None]] = []
-        self._hostname: str = ""
-        self._ip: str = ""
-
-    @property
-    def node_info(self) -> NodeInfo:
-        node_info = get_node_info()
-        assert node_info is not None
-        return node_info
-
-    def mock_file(self: NodeMockBase, fname: str, content: str) -> NodeMockBase:
-        self._files[fname] = content
-        return self
-
-    def mock_command_stdout(self: NodeMockBase, cmd: str, stdout: bytes | str) -> NodeMockBase:
-        self._stdout[cmd] = stdout
-        return self
-
-    def mock_http_response(self: NodeMockBase, method: str, url: str, response: Dict[str, Any]) -> NodeMockBase:
-        self._requests.append((method, url, response))
-        return self
-
-    def mock_hostname(self, hostname: str) -> NodeMockBase:
-        self._hostname = hostname
-        return self
-
-    def mock_ip(self, ip: str) -> NodeMockBase:
-        self._ip = ip
-        return self
-
-    def add_context(
-        self: NodeMockBase, ctx: ContextManager[Any], fn: Callable[[Any], None] | None = None
-    ) -> NodeMockBase:
-        self._contexts.append((ctx, fn))
-        return self
-
-    def _mock_file_open(self, fname: str) -> Any:
-        if fname not in self._files:
-            raise FileNotFoundError(f"File {fname} not found in mock")
-        return mock_open(read_data=self._files[fname]).return_value
-
-    async def _mock_create_subprocess_stdout(self, *args: Any, **kwargs: Any) -> Mock:
-        mock_proc = Mock()
-
-        async def mock_communicate() -> Tuple[bytes | str, bytes]:
-            cmd = args[0]
-            if cmd not in self._stdout:
-                raise Exception(f"Command '{cmd}' not found in mock")
-            return self._stdout[cmd], b""
-
-        mock_proc.communicate.side_effect = mock_communicate
-        return mock_proc
-
-    def _mock_subprocess_run_stdout(self, *args: Any, **kwargs: Any) -> Mock:
-        mock = Mock()
-        cmd = " ".join(args[0])
-        mock.stdout = self._stdout[cmd]
-        return mock
-
-    def _mock_http_response(self, mock: Any) -> None:
-        for method, url, response in self._requests:
-            mock.request(method, url, **response)
-
-    def _mock_local_ip(self, *args: Any, **kwargs: Any) -> Mock:
-        mock = Mock()
-        mock.getsockname.return_value = (self._ip, 0)
-        return mock
-
-    def __enter__(self: NodeMockBase) -> NodeMockBase:
-        self.add_context(
-            patch(
-                "builtins.open",
-                lambda fname, *args: self._mock_file_open(str(fname)),
-            )
-        )
-        self.add_context(
-            patch(
-                "asyncio.create_subprocess_shell",
-                self._mock_create_subprocess_stdout,
-            )
-        )
-        self.add_context(
-            patch(
-                "subprocess.run",
-                self._mock_subprocess_run_stdout,
-            ),
-        )
-        self.add_context(Mocker(), self._mock_http_response)
-
-        if self._hostname:
-            self.add_context(
-                patch(
-                    "os.uname",
-                    return_value=(
-                        "Linux",
-                        self._hostname,
-                        "5.15.0-79-generic",
-                        "#86-Ubuntu SMP Mon Jul 10 16:07:21 UTC 2023",
-                        "x86_64",
-                    ),
-                )
-            )
-
-        if self._ip:
-            self.add_context(
-                patch(
-                    "socket.socket",
-                    self._mock_local_ip,
-                )
-            )
-
-        for ctx, fn in self._contexts:
-            value = ctx.__enter__()
-            if fn is not None:
-                fn(value)
-        return self
-
-    def __exit__(self, exc_type: Any, exc_value: Any, exc_tb: Any) -> None:
-        for ctx, _ in self._contexts:
-            ctx.__exit__(None, None, None)
